completed project
biggest learning is to remember pi value (for Radar) should between +- 3.14159

+++ src/main.cpp
@@ -74,7 +74,7 @@ int main(int argc, char* argv[]) {
     MeasurementPackage meas_package;
     GroundTruthPackage gt_package;
     istringstream iss(line);
-    long long timestamp;
+    long timestamp;
 
     // reads first element from the current line
     iss >> sensor_type;
@@ -136,6 +136,7 @@ int main(int argc, char* argv[]) {
   for (size_t k = 0; k < N; ++k) {
     // start filtering from the second frame (the speed is unknown in the first
     // frame)
+
     fusionEKF.ProcessMeasurement(measurement_pack_list[k]);
 
     // output the estimation
